removed plugins and configs from main. added cli (start)
