+++ backend/src/cli/create_admin.py
@@ -1,0 +1,5 @@
+import click
+
+
+@click.command()
+async def create_admin() -> ...: ...
